Verify unspent key images when refreshing
Currently it is possible to lie to the SN network about your stake
balance by lying about your key images (for example, by just sending the
first couple key images but then stopping).  For example, with a balance
of 10K (plus a bit extra for fees) you could send yourself 10K 26 times,
then hack your supernode to only send the first key image (i.e. for the
intial 10K deposit) and have a supernode that is included in the auth
sample as a T4.  (There is a TODO in the code talking about verifying
the balance the SN sends, but this would provide no security: it would
take only slightly more work to make the announced balance agree with
the fake balance).

The main issue is that the wallet balance (via `wallet->balance()`)
can't be trusted because it's impossible to know if the remote SN sent
us all of its key images or not (we could just not have some because new
transactions have been received since the last announce)--and if it
didn't sent all of them, we will have some inputs that the wallet thinks
are unspent even if they are actually spent.

What needs to be done instead is to query only the verified-unspent
balance when determining a supernode stake, which this commit does.  In
the example above, the SN can cut the key image list at any point, but
now SNs simply won't recognize any desposits for which key images
haven't been sent.

There is a second issue, also fixed here, that a pool could broadcast a
legitimate stake (say 250K), then move the 250K to another wallet and
set that up as another supernode, but use a supernode hack that doesn't
announce relevant key images after the transfer.  The issue here is
that, even with the above changes, supernodes don't rescan the spent
status of transactions, and so with a dishonest SN, other SNs will never
learn that the transaction has spent (and thus the stake is no longer
actually staked).  The fix applied here is to force a rescan of unspent
key images when refreshing SN wallets.