Increase promise queue size to the next power of 2
For some reason the PromiseQueue produces a fatal error if given a queue
size that isn't a power of 2; passing it
`std::thread::hardware_concurrency()` thus breaks graft-ng on systems
with non-power-of-2 cores (e.g. on my 6c/12t server).  It was already a
reported problem on a single-core system.